+++ src/IngameIMEImpl.cpp
@@ -1,88 +1,10 @@
 #include <stdarg.h>
 
-#include "IngameIME.hpp"
+#include "ImmCompositionImpl.hpp"
+#include "ImmIngameIMEImpl.hpp"
 
-#include "CompositionImpl.hpp"
-#include "InputProcessorImpl.hpp"
-
-namespace libtf {
-    class GlobalImpl : public IngameIME::Global {
-      private:
-        DWORD cookieComp{TF_INVALID_COOKIE};
-        DWORD cookieProc{TF_INVALID_COOKIE};
-
-      protected:
-        ComPtr<InputProcessorHandler> handler;
-        friend class IngameIME::Global;
-
-      protected:
-        GlobalImpl()
-        {
-            COM_HR_BEGIN(S_OK);
-
-            handler = new InputProcessorHandler();
-
-            ComQIPtr<ITfSource> source(IID_ITfSource, handler->mode);
-            CHECK_HR(source->AdviseSink(
-                IID_ITfCompartmentEventSink, static_cast<ITfCompartmentEventSink*>(handler.get()), &cookieComp));
-
-            source = handler->compMgr;
-            CHECK_HR(source->AdviseSink(IID_ITfInputProcessorProfileActivationSink,
-                                        static_cast<ITfInputProcessorProfileActivationSink*>(handler.get()),
-                                        &cookieProc));
-
-            COM_HR_END();
-            COM_HR_THR();
-        }
-
-        ~GlobalImpl()
-        {
-            if (cookieProc != TF_INVALID_COOKIE) {
-                ComQIPtr<ITfSource> source(IID_ITfSource, handler);
-                source->UnadviseSink(cookieProc);
-                cookieProc = TF_INVALID_COOKIE;
-            }
-
-            if (cookieComp != TF_INVALID_COOKIE) {
-                ComQIPtr<ITfSource> source(IID_ITfSource, handler->mode);
-                source->UnadviseSink(cookieComp);
-                cookieComp = TF_INVALID_COOKIE;
-            }
-        }
-
-      public:
-        /**
-         * @brief Get Active InputProcessor
-         *
-         * @return std::shared_ptr<InputProcessor>
-         */
-        virtual std::shared_ptr<const IngameIME::InputProcessor> getActiveInputProcessor() const override
-        {
-            return IngameIME::InputProcessorImpl::getActiveInputProcessor();
-        }
-
-        /**
-         * @brief Get system availiable InputProcessor
-         *
-         * @return std::list<std::shared_ptr<InputProcessor>>
-         */
-        virtual std::list<std::shared_ptr<const IngameIME::InputProcessor>> getInputProcessors() const override
-        {
-            return IngameIME::InputProcessorImpl::getInputProcessors();
-        }
-
-        /**
-         * @brief Get the InputContext object
-         *
-         * @param hWnd the window to create InputContext
-         * @return std::shared_ptr<InputContext>
-         */
-        virtual std::shared_ptr<IngameIME::InputContext> getInputContext(void* hWnd, ...) override
-        {
-            return std::make_shared<InputContextImpl>(reinterpret_cast<HWND>(hWnd));
-        }
-    };
-}// namespace libtf
+#include "TfCompositionImpl.hpp"
+#include "TfIngameIMEImpl.hpp"
 
 libtf::InputContextImpl::InputContextImpl(HWND hWnd)
 {
@@ -113,49 +35,6 @@ libtf::InputContextImpl::InputContextImpl(HWND hWnd)
     COM_HR_THR();
 }
 
-namespace libimm {
-    class GlobalImpl : public IngameIME::Global {
-      public:
-        /**
-         * @brief Get Active InputProcessor
-         *
-         * @return std::shared_ptr<InputProcessor>
-         */
-        virtual std::shared_ptr<const IngameIME::InputProcessor> getActiveInputProcessor() const override
-        {
-            return IngameIME::InputProcessorImpl::getActiveInputProcessor();
-        }
-
-        /**
-         * @brief Get system availiable InputProcessor
-         *
-         * @return std::list<std::shared_ptr<InputProcessor>>
-         */
-        virtual std::list<std::shared_ptr<const IngameIME::InputProcessor>> getInputProcessors() const override
-        {
-            return IngameIME::InputProcessorImpl::getInputProcessors();
-        }
-
-        /**
-         * @brief Get the InputContext object
-         *
-         * @return std::shared_ptr<InputContext>
-         */
-        virtual std::shared_ptr<IngameIME::InputContext> getInputContext(void* hWnd, ...) override
-        {
-            std::shared_ptr<InputContextImpl> ctx;
-
-            auto iter = InputContextImpl::InputCtxMap.find(reinterpret_cast<HWND>(hWnd));
-            if (iter == InputContextImpl::InputCtxMap.end() || !(ctx = (*iter).second.lock())) {
-                ctx = std::make_shared<InputContextImpl>(reinterpret_cast<HWND>(hWnd));
-
-                InputContextImpl::InputCtxMap[reinterpret_cast<HWND>(hWnd)] = ctx;
-            }
-            return ctx;
-        }
-    };
-}// namespace libimm
-
 libimm::InputContextImpl::InputContextImpl(HWND hWnd) : hWnd(hWnd)
 {
     comp = std::make_shared<CompositionImpl>(this);
@@ -165,113 +44,6 @@ libimm::InputContextImpl::InputContextImpl(HWND hWnd) : hWnd(hWnd)
     ctx = ImmAssociateContext(hWnd, NULL);
 
     prevProc = (WNDPROC)SetWindowLongPtr(hWnd, GWLP_WNDPROC, (LONG_PTR)InputContextImpl::WndProc);
-}
-
-std::list<std::wstring> libimm::InputContextImpl::getInputModes()
-{
-    DWORD mode;
-    ImmGetConversionStatus(ctx, &mode, NULL);
-
-    auto activeProc = IngameIME::InputProcessorImpl::getActiveInputProcessor();
-
-    std::list<std::wstring> modes;
-    if (activeProc->type == IngameIME::InputProcessorType::KeyboardLayout)
-        modes.push_back(L"AlphaNumeric");
-    else {
-        if (mode & IME_CMODE_NATIVE) {
-            modes.push_back(L"Native");
-
-            if (activeProc->isJap)
-                if (mode & IME_CMODE_KATAKANA)
-                    modes.push_back(L"Katakana");
-                else
-                    modes.push_back(L"Hiragana");
-        }
-        else
-            modes.push_back(L"AlphaNumeric");
-
-        if (mode & IME_CMODE_FULLSHAPE)
-            modes.push_back(L"FullShape");
-        else
-            modes.push_back(L"HalfShape");
-    }
-
-    return modes;
-}
-
-IngameIME::InputProcessorContext libimm::InputContextImpl::getInputProcCtx()
-{
-    IngameIME::InputProcessorContext result;
-
-    result.proc  = IngameIME::InputProcessorImpl::getActiveInputProcessor();
-    result.modes = getInputModes();
-
-    return result;
-}
-
-LRESULT libimm::InputContextImpl::WndProc(HWND hWnd, UINT msg, WPARAM wparam, LPARAM lparam)
-{
-    auto iter = InputCtxMap.find(hWnd);
-
-    std::shared_ptr<InputContextImpl> inputCtx;
-    if (iter != InputCtxMap.end() && (inputCtx = (*iter).second.lock())) {
-        switch (msg) {
-            case WM_INPUTLANGCHANGE:
-                IngameIME::Global::getInstance().runCallback(IngameIME::InputProcessorState::FullUpdate,
-                                                             inputCtx->getInputProcCtx());
-                break;
-            case WM_IME_SETCONTEXT:
-                // We should always hide Composition Window to make the PreEditCallback for work
-                lparam &= ~ISC_SHOWUICOMPOSITIONWINDOW;
-
-                if (inputCtx->fullscreen) {
-                    // Hide Candidate Window
-                    lparam &= ~ISC_SHOWUICANDIDATEWINDOW;
-                }
-                break;
-            case WM_IME_STARTCOMPOSITION:
-                inputCtx->comp->IngameIME::PreEditCallbackHolder::runCallback(IngameIME::CompositionState::Begin,
-                                                                              nullptr);
-                return true;
-            case WM_IME_COMPOSITION:
-                if (lparam & (GCS_COMPSTR | GCS_CURSORPOS)) inputCtx->procPreEdit();
-                if (lparam & GCS_RESULTSTR) inputCtx->procCommit();
-
-                if (!inputCtx->fullscreen) inputCtx->procPreEditRect();
-
-                // when lparam == 0 that means current Composition has been canceled
-                if (lparam) return true;
-            case WM_IME_ENDCOMPOSITION:
-                inputCtx->comp->IngameIME::PreEditCallbackHolder::runCallback(IngameIME::CompositionState::End,
-                                                                              nullptr);
-                inputCtx->comp->IngameIME::CandidateListCallbackHolder::runCallback(IngameIME::CandidateListState::End,
-                                                                                    nullptr);
-                return true;
-            case WM_IME_NOTIFY:
-                if (inputCtx->fullscreen) switch (wparam) {
-                        case IMN_OPENCANDIDATE:
-                            inputCtx->comp->IngameIME::CandidateListCallbackHolder::runCallback(
-                                IngameIME::CandidateListState::Begin, nullptr);
-                            return true;
-                        case IMN_CHANGECANDIDATE: inputCtx->procCand(); return true;
-                        case IMN_CLOSECANDIDATE:
-                            inputCtx->comp->IngameIME::CandidateListCallbackHolder::runCallback(
-                                IngameIME::CandidateListState::End, nullptr);
-                            return true;
-                        default: break;
-                    }
-                if (wparam == IMN_SETCONVERSIONMODE) {
-                    IngameIME::Global::getInstance().runCallback(IngameIME::InputProcessorState::InputModeUpdate,
-                                                                 inputCtx->getInputProcCtx());
-                }
-                break;
-            case WM_IME_CHAR:
-                // Commit text already handled at WM_IME_COMPOSITION
-                return true;
-            default: return CallWindowProcW(inputCtx->prevProc, hWnd, msg, wparam, lparam);
-        }
-    }
-    return DefWindowProcW(hWnd, msg, wparam, lparam);
 }
 
 #include <versionhelpers.h>
