+++ src/include/common/InputContextImpl.hpp
@@ -1,0 +1,22 @@
+#pragma once
+#include <windows.h>
+
+#include "InputContext.hpp"
+
+namespace IngameIME {
+    struct InternalRect : public IngameIME::PreEditRect
+    {
+        InternalRect() = default;
+
+        operator RECT() noexcept
+        {
+            RECT rect;
+            rect.left   = this->left;
+            rect.top    = this->top;
+            rect.right  = this->right;
+            rect.bottom = this->bottom;
+
+            return rect;
+        }
+    };
+}// namespace IngameIME
