+++ src/include/tf/ComBSTR.hpp
