+++ src/include/tf/ComObjectBase.hpp
